changing behavior for rows with too few columns
Current behavior is to not insert rows if the number of columns in line
does not match number of columns in header.  Changing this to exclude
rows only if the number of columns in line is greater than number in
header.  For lines with less columns than header, will pad the end of
the line with enough Nils to make number of columns equal.

Reason for this is that some CSV files do not have commas at the end
of the file for null values and I want to insert these rows anyways,
instead of rejecting them.